pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "async-context-engine"
version = "4.0.0"
description = "Plug-and-play async context re-entry for LangGraph agents"
requires-python = ">=3.11"
license = "MIT"
readme = "README.md"
authors = [
    { name = "Sagar Mainkar" },
]
keywords = [
    "langgraph",
    "langchain",
    "async",
    "agents",
    "llm",
    "context-reentry",
    "background-tasks",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "langgraph>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0",
]
examples = [
    "langchain-ollama>=1.0",
    "python-dotenv>=1.0",
    "rich>=14.0",
]

[project.urls]
Homepage = "https://github.com/sagarmainkar/async-context-engine"
Repository = "https://github.com/sagarmainkar/async-context-engine"
Documentation = "https://github.com/sagarmainkar/async-context-engine/blob/main/docs/guide.md"
Issues = "https://github.com/sagarmainkar/async-context-engine/issues"

[tool.setuptools.packages.find]
include = ["async_context_engine*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
